Added exception handling methods in SystemOwnerServiceImpl and SpecialistServiceImpl. Created custom exception classes for these services and defined corresponding error messages. Also, added a new method to SpecialistRepository.

+++ src/main/java/org/example/clinicservice/service/impl/SystemOwnerServiceImpl.java
@@ -2,12 +2,17 @@ package org.example.clinicservice.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.example.clinicservice.entity.SystemOwner;
-import org.example.clinicservice.repository.SpecialistRepository;
+import org.example.clinicservice.exceptions.userExeptions.EmailNotFoundExсeption;
+import org.example.clinicservice.exceptions.ErrorMessage;
+import org.example.clinicservice.exceptions.userExeptions.InvalidIdException;
+import org.example.clinicservice.exceptions.systemOwnerExceptions.SystemOwnerExistsException;
+import org.example.clinicservice.exceptions.systemOwnerExceptions.SystemOwnerNotFoundException;
+import org.example.clinicservice.exceptions.userExeptions.InvalidPhoneNumberException;
+import org.example.clinicservice.exceptions.userExeptions.PhoneNumberNotFoundException;
 import org.example.clinicservice.repository.SystemOwnerRepository;
 import org.example.clinicservice.service.interfeces.SystemOwnerService;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -16,45 +21,68 @@ public class SystemOwnerServiceImpl implements SystemOwnerService {
 
     private final SystemOwnerRepository systemOwnerRepository;
 
+
     @Override
     public SystemOwner getSystemOwnerById(UUID ownerId) {
 
-        return systemOwnerRepository.findByOwnerId(ownerId);
+        try {
+            SystemOwner systemOwner = systemOwnerRepository.findByOwnerId(ownerId);
+            if(systemOwner == null){
+                throw new SystemOwnerNotFoundException(ErrorMessage.SYSTEM_OWNER_NOT_FOUND);
+            }
+            return systemOwner;
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while retrieving SystemOwner with id " + ownerId, e);
+        }
     }
 
     @Override
     public SystemOwner getSystemOwnerByEmail(String email) {
-
-        return systemOwnerRepository.findByEmail(email);
+        try {
+            SystemOwner systemOwner = systemOwnerRepository.findByEmail(email);
+            if(systemOwner == null){
+                throw new EmailNotFoundExсeption(ErrorMessage.EMAIL_DOES_NOT_EXIST);
+            }
+            return systemOwner;
+        } catch (Exception e) {
+            throw new RuntimeException("Error occurred while searching for system owner by email: " + email, e);
+        }
     }
 
     @Override
-    public SystemOwner getSystemOwnerByPhoneNumber(String phoneNumber) {
+    public SystemOwner findByPhoneNumber(String phoneNumber) {
+        try {
+            SystemOwner systemOwner = systemOwnerRepository.findByPhoneNumber(phoneNumber);
 
-        return systemOwnerRepository.findByPhoneNumber(phoneNumber);
-    }
+            if (systemOwner == null) {
+                throw new PhoneNumberNotFoundException(ErrorMessage.PHONE_NUMBER_DOES_NOT_EXIST);
+            }
 
-    @Override
-    public List<SystemOwner> getSystemOwnerByFirstNameAndLastName(String firstName, String lastName) {
-
-        return systemOwnerRepository.findByFirstNameAndLastName(firstName, lastName);
-    }
-
-    @Override
-    public List<SystemOwner> getSystemOwnersByEmailContaining(String email) {
-
-        return systemOwnerRepository.findByEmailContaining(email);
+            return systemOwner;
+        } catch (IllegalArgumentException e) {
+            throw new InvalidPhoneNumberException("Invalid phone number provided: " + e);
+        }
     }
 
     @Override
     public void saveSystemOwner(SystemOwner systemOwner) {
-
+        if (systemOwnerRepository.existsById(systemOwner.getOwnerId())) {
+            throw new SystemOwnerExistsException(ErrorMessage.SYSTEM_OWNER_WITH_ID_EXIST);
+        }
         systemOwnerRepository.save(systemOwner);
     }
 
     @Override
     public void deleteSystemOwner(UUID ownerId) {
 
-        systemOwnerRepository.deleteById(ownerId);
+        try {
+            if (!systemOwnerRepository.existsById(ownerId)) {
+                throw new SystemOwnerNotFoundException(ErrorMessage.USER_WITH_ID_NOT_EXIST);
+            }
+            systemOwnerRepository.deleteById(ownerId);
+        } catch (IllegalArgumentException e) {
+            throw new InvalidIdException(ErrorMessage.INVALID_USER_ID);
+        }
     }
 }
+
